Update culling to be done with double precision.

+++ modules/globebrowsing/rendering/culling.h
@@ -67,7 +67,7 @@ namespace openspace {
         */
         static bool isVisible(
             const RenderData& data,
-            const vec3& point);
+            const dvec3& point);
 
 
         /**
@@ -93,17 +93,17 @@ namespace openspace {
         what is considered to be inside the view frustrum.
         */
         static PointLocation testPoint(
-            const glm::vec2& pointScreenSpace,
-            const glm::vec2& marginScreenSpace = vec2(0));
+            const glm::dvec2& pointScreenSpace,
+            const glm::dvec2& marginScreenSpace = vec2(0));
 
         static bool testPoint(
-            const glm::vec3& pointScreenSpace,
-            const glm::vec3& marginScreenSpace = vec3(0));
+            const glm::dvec3& pointScreenSpace,
+            const glm::dvec3& marginScreenSpace = vec3(0));
 
 
-        static glm::vec2 transformToScreenSpace(
-            const vec3& point,
-            const mat4x4& modelViewProjection);
+        static glm::dvec2 transformToScreenSpace(
+            const dvec3& point,
+            const dmat4x4& modelViewProjection);
 
     };
